# sumi-daemon/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sumi-daemon | v1",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:27381/"
    }
  ],
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "SSE 事件订阅：书库变更推送。帧格式 `event: <事件名>` + `data: <JSON 载荷>`；\n全部事件名与载荷契约见 core::events（事件名以常量集中定义，即持久契约）。\n消费跟不上（lagged）或总线关闭即断开；重连后须以 item/skeleton + folder/list 全量对齐",
        "operationId": "events",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "访问 token（EventSource 无法设置请求头）",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream 长连接",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/item/content": {
      "get": {
        "tags": [
          "item"
        ],
        "summary": "`GET /api/v1/item/content?id=`：归一化阅读内容（按需转换，结果入派生缓存，immutable）",
        "operationId": "content",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML 或 UTF-8 文本"
          }
        }
      }
    },
...
    "/api/v1/item/cover": {
      "get": {
        "tags": [
          "item"
        ],
        "summary": "`GET /api/v1/item/cover?id=`：封面（自定义优先于自动提取；自动封面四分支响应契约见 API 文档）",
        "operationId": "cover_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "封面 webp 字节"
          }
        }
      },
      "put": {
        "tags": [
          "item"
        ],
        "summary": "`PUT /api/v1/item/cover`：设置自定义封面（解码校验→缩入 1024→webp→.sumi/covers/）",
        "operationId": "cover_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "item"
        ],
        "summary": "`DELETE /api/v1/item/cover?id=`：移除自定义封面（幂等；回退自动提取链）",
        "operationId": "cover_delete",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/item/file": {
      "get": {
        "tags": [
          "item"
        ],
        "summary": "`GET /api/v1/item/file?id=`：原书文件字节（支持 Range；Cache-Control immutable）",
        "operationId": "file",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "原文件字节"
          }
        }
      }
    },
    "/api/v1/item/list": {
      "post": {
        "tags": [
          "item"
        ],
        "summary": "`POST /api/v1/item/list`：查询 item 列表（过滤条件复杂故 POST；全部 AND）",
        "operationId": "list",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/item/resource": {
      "get": {
        "tags": [
          "item"
        ],
        "summary": "`GET /api/v1/item/resource?id=&path=`：归一化 HTML 引用的书内嵌资源（epub/docx zip 按需解包）",
        "operationId": "resource",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "资源字节"
          }
        }
      }
    },
...
    "/api/v1/item/skeleton": {
      "post": {
        "tags": [
          "item"
        ],
        "summary": "`POST /api/v1/item/skeleton`：全量布局骨架（与 list 同序，不分页，只返回布局所需最低字段）",
        "operationId": "skeleton",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/item/toc": {
      "get": {
        "tags": [
          "item"
        ],
        "summary": "`GET /api/v1/item/toc?id=`：阅读目录（章节树；anchor 为 opaque 定位符）",
        "operationId": "toc",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoverPut": {
        "type": "object",
        "required": [
          "id",
          "img_base64"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "img_base64": {
            "type": "string"
          }
        }
      },
...
      "ListQuery": {
        "type": "object",
        "properties": {
          "annotation": {
            "type": [
              "string",
              "null"
            ]
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "categories_match": {
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "exclude_authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exclude_categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exclude_folders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exclude_tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ext": {
            "type": [
              "string",
              "null"
            ]
          },
          "folders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "folders_exact": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "in_trash": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "isbn": {
            "type": [
              "string",
              "null"
            ]
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "type": [
              "string",
              "null"
            ]
          },
          "limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "offset": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "order": {
            "type": [
              "string",
              "null"
            ]
          },
          "order_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ]
          },
          "read_status": {
            "type": [
              "string",
              "null"
            ]
          },
          "series": {
            "type": [
              "string",
              "null"
            ]
          },
          "star": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          },
          "without_authors": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "without_categories": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "without_tags": {
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}